Fix crash when converting `Variant` to incompatible `RequiredPtr<T>`
